Clean up RaceModel
Remove global driver list and replace with method to get new list
as needed. Adjust code in RaceModel accordingly.

Having a global driver list raises issues when accounting for
aspects such as DNFs. Instead of adding more code to Driver class
and bloating it, it is easier to have a local, modifiable list that
can be generated and accessed with a method call.